+++ QuadVertexData.cpp
@@ -15,7 +15,7 @@ const std::vector<QuadVertex> VERTICES =
 QuadVertexData::QuadVertexData()
 {
     indexBufferObject = new TransferBufferObject(VK_BUFFER_USAGE_INDEX_BUFFER_BIT);
-    indexCount = INDICES.size();
+    indexCount = static_cast<uint32_t>(INDICES.size());
     indexType = VK_INDEX_TYPE_UINT16;
     vertexBufferObject = new TransferBufferObject(VK_BUFFER_USAGE_VERTEX_BUFFER_BIT);
 }
